moved getNonNullHeader to generic class
