+++ src/main/java/org/gridsuite/loadflow/server/computation/service/AbstractComputationService.java
@@ -7,10 +7,12 @@
 package org.gridsuite.loadflow.server.computation.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.powsybl.commons.PowsyblException;
 import lombok.Getter;
 import org.gridsuite.loadflow.server.computation.repositories.ComputationResultRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.messaging.MessageHeaders;
 
 import java.util.List;
 import java.util.Objects;
@@ -64,5 +66,13 @@ public abstract class AbstractComputationService<R> {
 
     public void deleteResults() {
         resultRepository.deleteAll();
+    }
+
+    public static String getNonNullHeader(MessageHeaders headers, String name) {
+        String header = (String) headers.get(name);
+        if (header == null) {
+            throw new PowsyblException("Header '" + name + "' not found");
+        }
+        return header;
     }
 }
